First Scenario w/o Accusation
Can get a fact from talking to Waldorf and a fact from talking to the
cat. Can combine them and use the resulting fact to accuse the cat.
Accusation sequence begins but is not complete.

Can pick up an item (cat hair). Should be able to use it as optional
evidence in next commit.

Fixed bug with inventory screen index going out of range. Added sliding
movement to the player.

TO-DO: Finish accusation sequence. Understand how to properly trigger
cutscenes and animations.